Front-end/requirement.txt
Flask
youtube_transcript_api
google-generativeai
nltk
